Create SelfHostedOrganizationLicensesController
